+++ ayaka/core/database.py
@@ -11,8 +11,10 @@ from ..helpers import ensure_dir_exists, simple_async_wrap
 class AyakaDB:
     '''数据库'''
 
-    def __init__(self, name: str = "ayaka", _auto_regist: bool = True) -> None:
+    def __init__(self, name: str = "ayaka") -> None:
         self.name = name
+        db_dict[name] = self
+
         self.metadata = MetaData()
 
         class Model(SQLModel):
@@ -22,12 +24,20 @@ class AyakaDB:
             def __tablename__(cls) -> str:
                 return inflection.underscore(cls.__name__)
 
+            @classmethod
+            def get_session(cls, **kwargs):
+                '''创建一个新的数据库session
+
+                kwargs 请参考sqlmodel.Session'''
+                return self.get_session(**kwargs)
+
         self.SQLModel = Model
 
         class DBBase(Model):
             id: Optional[int] = Field(None, primary_key=True)
 
         self.EasyModel = DBBase
+        '''自带id'''
 
         class GroupDBBase(Model):
             group_id: str = Field(primary_key=True)
@@ -48,6 +58,7 @@ class AyakaDB:
                 return cls._get_or_create(group_id=group_id)
 
         self.GroupDBBase = GroupDBBase
+        '''自带group_id'''
 
         class UserDBBase(GroupDBBase):
             user_id: str = Field(primary_key=True)
@@ -57,9 +68,9 @@ class AyakaDB:
                 return cls._get_or_create(group_id=group_id, user_id=user_id)
 
         self.UserDBBase = UserDBBase
+        '''自带group_id、user_id'''
 
-        if _auto_regist:
-            get_adapter().on_startup(simple_async_wrap(self.init))
+        get_adapter().on_startup(simple_async_wrap(self.init))
 
     def init(self):
         '''初始化所有表'''
@@ -74,7 +85,6 @@ class AyakaDB:
 
         kwargs 请参考sqlmodel.Session'''
 
-        kwargs.setdefault("bind", self.engine)
         # 在数据库高频修改的情况下，autoflush可能会导致数据错误，例如多次相加只生效1次
         # 但减少了数据库崩溃报错的可能，并且提高数据库吞吐量
         kwargs.setdefault("autoflush", False)
@@ -84,4 +94,10 @@ class AyakaDB:
         return Session(**kwargs)
 
 
-ayaka_db = AyakaDB(_auto_regist=False)
+db_dict: dict[str, AyakaDB] = {}
+
+
+def get_db(name: str = "ayaka"):
+    if name not in db_dict:
+        db_dict[name] = AyakaDB(name)
+    return db_dict[name]